Several fixes to the bug triage dashboard to make it more useful
The top-bugs section now bail early if there isn't a top-<tool>-bug
found to avoid throwing.
Also, in the case of the inspector, it now searches for [dt-q] bugs
instead.

The intermittent bugs section now searches for [stockwell-needswork]
bugs instead.

And a minor CSS fix to make sure there's padding at the bottom of
page.